+++ Runtime/Timer.cs
@@ -10,27 +10,34 @@ namespace Scream.UniMO
     [System.Serializable]
     public class UnscaledTimer
     {
-        float timeSection;
-        float timer;
+
 #if UNITY_EDITOR
         [ReadOnly, SerializeField] float remain = 0f;
         [ReadOnly, SerializeField] bool bFinished = false;
 #endif
+        float timeSection = 0f;
+        float timer = 0f;
+        float remainTime = 0f;
+
+        /// <summary> timer is in pause state </summary>
+        public bool IsPausing { get; private set; } = false;
+
+        /// <summary>return the cd range from 0 to 1 0 means timer finished </summary>
+        public float Remain01 => Remain / timeSection;
 
         /// <summary>remaining time until the countdown end</summary>
         public float Remain
         {
             get
             {
-                float offset = timer - UnityEngine.Time.unscaledTime;
+                float offset = timer - Time.unscaledTime;
+                if (IsPausing)
+                    return remainTime;
                 if (offset >= 0f)
                     return offset;
                 return 0f;
             }
         }
-
-        /// <summary>return the cd range from 0 to 1 0 means timer finished </summary>
-        public float Remain01 => Remain / timeSection;
 
         /// <summary>if this countdown finished or not</summary>
         public bool IsFinished
@@ -38,30 +45,50 @@ namespace Scream.UniMO
             get
             {
 #if UNITY_EDITOR
-                bFinished = timer <= UnityEngine.Time.unscaledTime;
-                remain = this.Remain;
+                bFinished = timer <= Time.unscaledTime;
+                remain = Remain;
 #endif
-                return timer <= UnityEngine.Time.unscaledTime;
+                return timer <= Time.unscaledTime && !IsPausing;
             }
         }
 
         public UnscaledTimer(float timeSection = 0f, bool CanUseFirst = true)
         {
             this.timeSection = timeSection;
+            IsPausing = false;
             if (!CanUseFirst)
                 Reset();
             else
-                this.timer = 0f;
+                timer = 0f;
+        }
+
+        /// <summary>Pause this timer</summary>
+        public void Pause()
+        {
+            IsPausing = true;
+            remainTime = timer - Time.unscaledTime;
+        }
+
+        /// <summary>Resume this timer</summary>
+        public void Resume()
+        {
+            IsPausing = false;
+            timer = Time.unscaledTime + remainTime;
         }
 
         /// <summary>Reset countdown timer with default setting</summary>
-        public void Reset() => timer = UnityEngine.Time.unscaledTime + timeSection;
+        public void Reset()
+        {
+            timer = Time.unscaledTime + timeSection;
+            IsPausing = false;
+        }
 
         /// <summary>reset countdown timer with new timeSection</summary>
         public void Reset(float timeSection)
         {
             this.timeSection = timeSection;
-            timer = UnityEngine.Time.unscaledTime + timeSection;
+            timer = Time.unscaledTime + timeSection;
+            IsPausing = false;
         }
     }
 
@@ -70,26 +97,34 @@ namespace Scream.UniMO
     [System.Serializable]
     public class ScaledTimer
     {
-        float timeSection;
-        float timer;
+
 #if UNITY_EDITOR
         [ReadOnly, SerializeField] float remain = 0f;
         [ReadOnly, SerializeField] bool bFinished = false;
 #endif
+        float timeSection = 0f;
+        float timer = 0f;
+        float remainTime = 0f;
+
+        /// <summary> timer is in pause state </summary>
+        public bool IsPausing { get; private set; } = false;
+
+        /// <summary>return the cd range from 0 to 1 0 means timer finished </summary>
+        public float Remain01 => Remain / timeSection;
+
         /// <summary>remaining time until the countdown end</summary>
         public float Remain
         {
             get
             {
-                float offset = timer - UnityEngine.Time.time;
+                float offset = timer - Time.time;
+                if (IsPausing)
+                    return remainTime;
                 if (offset >= 0f)
                     return offset;
                 return 0f;
             }
         }
-
-        /// <summary>return the cd range from 0 to 1 0 means timer finished </summary>
-        public float Remain01 => Remain / timeSection;
 
         /// <summary>if this countdown finished or not</summary>
         public bool IsFinished
@@ -97,30 +132,50 @@ namespace Scream.UniMO
             get
             {
 #if UNITY_EDITOR
-                bFinished = timer <= UnityEngine.Time.time;
-                remain = this.Remain;
+                bFinished = timer <= Time.time;
+                remain = Remain;
 #endif
-                return timer <= UnityEngine.Time.time;
+                return timer <= Time.time && !IsPausing;
             }
         }
 
         public ScaledTimer(float timeSection = 0f, bool CanUseFirst = true)
         {
             this.timeSection = timeSection;
+            IsPausing = false;
             if (!CanUseFirst)
                 Reset();
             else
-                this.timer = 0f;
+                timer = 0f;
+        }
+
+        /// <summary>Pause this timer</summary>
+        public void Pause()
+        {
+            IsPausing = true;
+            remainTime = timer - Time.time;
+        }
+
+        /// <summary>Resume this timer</summary>
+        public void Resume()
+        {
+            IsPausing = false;
+            timer = Time.time + remainTime;
         }
 
         /// <summary>Reset countdown timer with default setting</summary>
-        public void Reset() => timer = UnityEngine.Time.time + timeSection;
+        public void Reset()
+        {
+            timer = Time.time + timeSection;
+            IsPausing = false;
+        }
 
         /// <summary>reset countdown timer with new timeSection</summary>
         public void Reset(float timeSection)
         {
             this.timeSection = timeSection;
-            timer = UnityEngine.Time.time + timeSection;
+            timer = Time.time + timeSection;
+            IsPausing = false;
         }
     }
 }