+++ CHANGELOG.md
@@ -1,8 +1,17 @@
-# 1.1.0
+## 1.2.0
+### Timer
+#### Add
+- `bool IsPausing` timer is in pause state
+- `Pause()` Pause this timer
+- `Resume()` Resume this timer
+#### Update
+- `bool IsFinished` will keep false when timer is in pause state
+
+## 1.1.0
 - Add JsonFileUtility
 
-# 1.0.2
+## 1.0.2
 - Fix Scream.UniMO.Editor Build Error
 
-# 1.0.1
-- Disable Scream.UniMO.Input.GamepadControler
+## 1.0.1
+- Disable Scream.UniMO.Input.GamepadController